save vrm in app context
