+++ src/common/context/index.tsx
@@ -1,3 +1,5 @@
-export { default } from "./context.menu";
+export { MenuProvider } from "./context.menu";
+export { VRMProvider } from "./context.vrm";
 
 export { MenuContext } from "./context.menu";
+export { VRMContext } from "./context.vrm";
