+++ src/pages/index.astro
@@ -8,17 +8,6 @@ import Composition from "3D/Canvas";
 <Layout>
   <Menu client:only="react" />
   <Cameras client:only="react" />
-  <p
-    class="absolute top-1/2 left-1/2 z-0"
-    style="transform:translate(-50%,-50%)"
-  >
-    Press Space
-  </p>
   <Composition client:only="react" />
 </Layout>
-<script>
-  // import { state } from "state";
-  // console.log(state.gameStarted);
-
-  // const;
-</script>
+<script></script>
